+++ app/components/primer/button_component.html.erb
@@ -1,4 +1,4 @@
-<%= render Primer::BaseButton.new(**@system_arguments) do -%>
+<%= render Primer::Beta::BaseButton.new(**@system_arguments) do -%>
   <%= leading_visual %><%= trimmed_content %><%= trailing_visual %><%= primer_octicon("triangle-down", ml: 2, mr: -1) if @dropdown %>
   <%= tooltip %>
 <% end -%>
