+++ static/audited_at.json
@@ -11,12 +11,12 @@
   "Primer::Alpha::Tooltip": "",
   "Primer::Alpha::UnderlineNav": "",
   "Primer::Alpha::UnderlinePanels": "",
-  "Primer::BaseButton": "",
   "Primer::BaseComponent": "",
   "Primer::Beta::AutoComplete": "",
   "Primer::Beta::AutoComplete::Item": "",
   "Primer::Beta::Avatar": "",
   "Primer::Beta::AvatarStack": "",
+  "Primer::Beta::BaseButton": "",
   "Primer::Beta::Blankslate": "",
   "Primer::Beta::Breadcrumbs": "",
   "Primer::Beta::Breadcrumbs::Item": "",
