+++ stories/primer/beta/base_button_stories.rb
@@ -1,16 +1,19 @@
 # frozen_string_literal: true
 
-class Primer::IconButtonStories < ViewComponent::Storybook::Stories
+require "primer/beta/base_button"
+
+class Primer::Beta::BaseButtonStories < ViewComponent::Storybook::Stories
   layout "storybook_preview"
 
-  story(:icon_button) do
+  story(:base) do
     controls do
-      aria(label: "Button label")
-      icon "star"
-      box false
+      block false
       select(:tag, Primer::Beta::BaseButton::TAG_OPTIONS, Primer::Beta::BaseButton::DEFAULT_TAG)
       select(:type, Primer::Beta::BaseButton::TYPE_OPTIONS, Primer::Beta::BaseButton::DEFAULT_TYPE)
-      select(:scheme, Primer::IconButton::SCHEME_OPTIONS, Primer::IconButton::DEFAULT_SCHEME)
+    end
+
+    content do
+      "Click me"
     end
   end
 end
