--- test/components/base_button_test.rb
@@ -1,41 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class PrimerBaseButtonTest < Minitest::Test
-  include Primer::ComponentTestHelpers
-
-  def test_renders_content
-    render_inline(Primer::BaseButton.new) { "content" }
-
-    assert_text("content")
-  end
-
-  def test_renders_a_without_button_role
-    render_inline(Primer::BaseButton.new(tag: :a)) { "content" }
-
-    assert_selector("a")
-    refute_selector("a[role='button']")
-    refute_selector("a[type]")
-  end
-
-  def test_renders_summary_without_button_role
-    render_inline(Primer::BaseButton.new(tag: :summary)) { "content" }
-
-    assert_selector("summary")
-    refute_selector("summary[role='button']")
-    refute_selector("summary[type]")
-  end
-
-  def test_renders_href
-    render_inline(Primer::BaseButton.new(href: "www.example.com")) { "content" }
-
-    assert_selector("button[href='www.example.com']")
-  end
-
-  def test_renders_button_block
-    render_inline(Primer::BaseButton.new(block: true)) { "content" }
-
-    assert_selector(".btn-block")
-  end
-end
